Move shimmering method to extension file
